universal.make: Copy-and-paste and delete bits of the "dist" variables and rules to make a new "native-dist" target which builds and scp()s, for example, salma-hayek-Linux-2005-09-13.tar.gz and then makes a symlink salma-hayek-Linux.tgz in the parent directory.
per-directory.make: Cygwin make contains a hack so that rules which have a target of "badger" are understood to mean "badger.exe" but tar, obviously, can contain no such hack.  So, now we pass $(EXECUTABLE) (indirectly) to tar, it needs to contain the ".exe".  That doesn't seem to break the normal build.