www/index.html: It's not entirely obvious how to run terminator from a Shortcut on the Windows desktop.  Getting rid of the "spurious" window also took a deal of research, despite the solution being easy.
I've attempted to reassure the reader that they're not losing, for example, any unhandled exception messages from the Java.  We don't make those messages easy to find.  We should give the path to the terminator log as well as the terminal log in the Show Info box.  We can't expect users to run with DEBUGGING_TERMINATOR=true, even if run like that.

It's so slow to start a new terminator on Windows that the Alt-N shortcut is worth mentioning.  We know that users don't discover the shortcut for themselves.