Although launcher.exe has been giving me terminator log output when terminator is run from a terminator/bash window and DEBUGGING_TERMINATOR=true, it hasn't been giving me the output when run via a Shortcut to ruby.exe.  That's even though ruby.exe is a console subsystem application (unlike rubyw.exe) and has a console window.
So I thought it was time to bite the bullet and build launcher.exe and launcherw.exe, following the ruby.exe and rubyw.exe convention.  (As opposed to perl.exe and wperl.exe - I don't think it would have taken me so long to discover a perlw.exe.)

per-directory.make: It turns out to be easy to build more than one executable.  Most of the changes are cosmetic pluralizations.  The one wrinkle was that there is no built-in rule for building %w.exe from %.o (as opposed to %.exe from %.o).  That actual led to less code: we can use the same rule for all linked targets - shared libraries and executables.

universal.make: Adapt to the pluralization and let per-directory.make decide when to apply the --subsystem flag.

With these changes, you'll see console output from terminator until I check-in my change to use launcherw when DEBUGGING_TERMINATOR isn't set.

There also seems to be a bug in my local-variables checking, meaning that your first build after this change might fail with an error like:

../salma-hayek/universal.make:484: *** makefile bug: local variable EXECUTABLE from scope launcher (with value "") was referred to in scope setsid.  Stop.

A subsequent make should succeed.  I'll look at this after checking in terminator.  One thing at a time...