bin/target-os.rb: A new script to cope with Cygwin's uname, which is CYGWIN_NT-5.1 on XP and CYGWIN_NT-5.0 on Windows 2000.
http://www.opengroup.org/onlinepubs/009695399/utilities/uname.html suggests that the output of uname with no options shouldn't include any version information.  Indeed, the cygwin version is absent, but it doesn't seem right to include the underlying OS version either.  Solaris, for example, doesn't tell you whether it's running on ix86 or SPARC without any options.

universal.make: 
per-directory.make: Use the new script to name the target directory and to select the compiler flags etc.  One day, we'll have to distinguish between different architectures running the same OS, but not today.

www/index.html: That means that the name of the native-dist has changed.

bin/fatbits: 
bin/javahpp.rb: Use the new script in all the Ruby scripts in preference to uname.  The scripts which care about Cygwin are all Ruby.  There are still some bash relics which I haven't converted.