src/e/ptextarea/PTextArea.java: Clicking in a PTextArea to position the caret or to give focus was clearing the X11 selection.  I obviously use the middle button to paste the X11 selection more than I'd thought because this was really annoying me.  If I clear Firefox's selection, it doesn't clear what it's put in the X11 selection.
I missed the almost duplicate code in copy() before.  If we use the length of the selected text instead of checking hasSelection, then we can definitely rely on getSelectedText's lock.  JComponent gives us a getToolkit which we should use in preference to the default toolkit, for one thing because it's less work.