universal.make: $(RM) isn't rm, it's rm -f.  That isn't my doing - it's built-in to make.  The --no-run-if-empty dates from a time when I either didn't know that $(RM) was the make convention or (more likely) had temporarily forgotten.  I should have removed the --no-run-if-empty when I switched to using $(RM) "everywhere".  I'm ambivalent about deviating from the $(RM) convention.
Contrary to the previous check-in comment, both Mac OS X and Solaris's rm(1) complain if given no arguments (at least in the versions I have easy access to).

I'd rather not contend with nested quotes and local bash variables and the previous author was right to be wary of using find(1) with a make variable.