src/e/ptextarea/PReverseSegmentIterator.java: Add a new segment iterator which iterates backwards.  This allows us to change the bracket matching code to...
src/e/ptextarea/PBracketUtilities.java: ...use that, in much the same way as the forward bracket matching.  I've got rid of the stopwatch stuff now, too.  This method is now fast.

src/e/ptextarea/PMatchingBracketHighlighter.java: Now that finding matching brackets is not slow, we don't need to claim it is, and we don't need to break early.  We can also elide the strangely separate 'if (highlights[0] == null)' into a chained else clause.

src/e/ptextarea/PMouseHandler.java: Again, stop moaning about the sale-style char sequence being hideously expensive.  It is, but we don't use it anymore.  In fact, I'm tempted to delete the thing.  While I was at it, I killed off a left-over call to PLineSegment.getCharSequence().