I've added -server to my terminator options to test Ed's theory that the server VM will give back memory to the operating system.  It doesn't seem to for me but I haven't investigated further than top(1).  In the meantime, I'd like to keep my Win32 build runnable.
native/all/launcher/launcher.cpp: Wanting to support -server makes it clear on what object most of the previously static methods belong.  Judging by the order of the help which java -help produces (perhaps not the best guide), the client/server choice might be just one that the Sun launcher supports: the 32/64 bit choice might also affect which JVM is chosen.

While writing this, I discovered that Cygwin's boost support is a bit of a pain.  It's quite a large download (minutes) and then, seemingly, you manually have to create a symlink in /usr/include to the version you'd like to use.  Or possibly it's just currently broken - tic(1) seems to be, as does cygwin.com.