per-directory.make: The command for making the file list is about to get significantly more complicated.  Continue my silent demonstration that rules aren't where the complexity lies in any sizeable make system by losing this command to the place in universal.make.  That's where we know about all the other native targets that need including in the installer.  As long as we don't include them in dependencies here, we won't get into a circularity problem.
I'm about to add dependencies on the other native targets.  We don't want to pass these - ie the totality of $^ - directly to the WiX linker.  The idiom I've developed here to safely refer to per-directory local variables using target-specific variables encouraged me to abstract-out the duplicate WiX linker rule, in order that I could then solve this problem by using $(WIX_OBJECTS).  Then I got stuck: we can't correctly refer to both the local variable $(WIX_OBJECTS) and the automatic variable $@.  If we evaluate too early, $@ won't be set.  If we evaluate too late, $(WIX_OBJECTS) won't be set any more (and the universal.make local variable support will enforce this).  So I had to go back to using $(filter) on $^.  It then seemed like a backwards step to re-add the duplication of the rule even though, unusually, it has to be (and can safely be) a deferred evaluation assignment.

universal.make: Filter the installers out of the list of all native targets.  Add all the native targets except the installer(s) to the list of components for the installer compiler to bundle up.  Make the installer depend on all the other native targets.  Blimey, that was surprisingly easy: the first idea I came up worked.

This produces the first (I think) complete installer for terminator.  There's no shortcut for it yet, so I haven't made any effort to test it, bar browsing the installed files to see that everything looks present.