native/Cygwin/WiX/terminator.wxs: *What* a palaver.  I couldn't find any examples of anyone setting a shortcut to point to file they didn't install.  I had to go back and read through more of the tutorial.  FileSearch had to be the answer but what a hairy way to specify a path!  And this isn't even what I really want to do, which is:
`cygpath --windows rubyw.exe`

bin/terminator: Oh no, we're installed in a directory with a space in the name.  Can this really be the only problem line?  It seems to be, but I fear this isn't a maintainable way to develop.  I can see my having to come along adding random quotes when someone else breaks the Windows build.  It would be a more sustainable idea to use cygpath in a similar way to that used by universal.make, to convert the salma_hayek and terminator_home variables to forms without spaces.  Who said DOS names are useless?

Anyway, this gets us the first really fully working, useful Windows install.  Woo hoo!