native/Cygwin/WiX/terminator.wxs: A blank icon is surprisingly hard to find.  Use the Cygwin icon, as we invoke the Cygwin shell by default.
I thought I was being clever with the change to use $(env.SystemDrive) for finding the Cygwin installation's rubyw.exe.  It was only when I used SystemDrive again to find the icon and then realized that it's bundling the icon into the .msi that I realized it's evaluating SystemDrive on my machine, not on the target machine.