src/terminator/terminal/TerminalControl.java: Defer the decision on holding the window open when the child dies with a non-zero status...
src/terminator/view/JTerminalPane.java: ... so that shell windows can behave differently to, say, telnet windows.  If telnet fails, which happens quite often, then the user won't get to see the error message unless we hold the window open.  bash, on the other hand, won't fail (except in pathological cases).  It will often, however, return a non-zero exit status.  It does this when the last command it ran failed.  The user will have already seen the failure.  It's boring to require the user to manually close the window in this common case.