src/e/edit/Workspace.java: write each window's y-coordinate to the XML saved state, plus support for opening a window at a given y-coordinate.
src/e/edit/EColumn.java: support for adding a window at a given y-coordinate and overriding the heuristic placement.

src/e/edit/Edit.java: read XML saved state that may or may not contain y-coordinates. Private API (the InitialFile class) is used to carry around the filename and (optional) y-coordinate. This might -- if we can find a decent name -- be a good candidate for splitting out and moving more of the "open" logic into. For now, it's a crufty implementation detail complete with overloading and non-private fields.