src/e/edit/Advisor.java: A better way to ensure there's only one task running at a time. If the researchers took a long time it might be worth killing them, but they shouldn't be taking a long time.
src/e/edit/TagsUpdater.java: Do the same, to make sure only one tags panel is
updated at a time. If the Executor executes the tasks in the order it gets them,
this should help to ensure that when starting Edit with some files open, the
tags panel contents correspond to the ETextWindow opened last, rather than the
one whose TagsUpdater finished last.