Use the majority of the saving code to test whether a new charset encoding will cope with the file contents.
src/e/ptextarea/PTextBuffer.java: Split writeToFile in two, making a writeToStream.  writeToFile keeps the file-specific parts.  It opens the file, so it closes it (though OutputStreamWriter is a Closeable, so this could be shared).  attemptEncoding works like attemptDecoding - it switches the buffer's charset property if the encoding completes without throwing an exception.

makeReportingCharsetDecoder looks like it may have been written by someone under a misapprehension as to the default behavior of CharsetDecoder.  Perhaps it was just defensive or an attempt to reveal the intention but the lack of a comment suggests that it was just a mistake.  The OutputStreamWriter behavior coupled with its lack of documentation do leave one with the incorrect impression that the default is the opposite.

If this check-in changes the behavior of saving (or loading), that's unintentional.  It's only supposed to provide an additional method for the benefit of Edit's File Properties dialog.

src/e/util/NullOutputStream.java: Make it easy to process output, to see if the processing will work, without saving the result.