src/e/edit/FilePropertiesAction.java: Allow the user to change the charset encoding.  The input UI could more usefully be a combo box with a list containing ASCII (for programs which must be ASCII), ISO-8859-1 (for the majority of legacy stuff), UTF-8 (for almost everything) and perhaps Windows code pages and something relating to little-endian and big-endian UCS-2 (or whatever it's called).  Text's adequate as an interim, though, given that we validate it.  A more interesting question is what to make the output UI: how to present encoding failures.  Being able to step through the encoding failures as if they were Find matches might be useful.  That suggests that the attemptEncoding method signature would want to change but, given the current lack of UI, the current signature seems reasonable at the moment.
My motivation here was to save octopus/sib/fc/Oui.cpp in UTF-8 so that it would (as it turned out) compile cleanly with a new version of icc.  I didn't know about iconv(1) at the time and Markus Kuhn's normally helpful page just left me drowning in esoterica.  I did later try pasting the 400KiB source file into a new buffer but that locked up the EDT for, say, 15 minutes, re-indenting the pasted version, before I gave up on it and killed Edit.

Having written the code, though, it seems odd not to let the user change the charset property, at least now we can determine that it's a non-lossy change.