src/e/edit/FileIgnorer.java: Pull the file and directory ignoring code out of salma-hayek.  It currently has two users in Edit - the workspace scanner and something to do with the JavaDoc stuff - and none elsewhere.  The FileIgnorer is now an object rather than a bunch of static methods.  It's constructor takes the root directory, so the class has an opportunity to load settings particular to a workspace.  I've yet to use the parameter: small steps.
src/e/edit/JavaDoc.java: Change to use FileIgnorer instead of FileUtilities.  The lifetime of the FileIgnorer is currently one call to findPackagesIn.  I don't know if that's the best choice.

src/e/edit/Workspace.java: Change to use FileIgnorer instead of FileUtilities.  Remove the nameEndsWithOneOf(ignoredExtensions) check as FileUtilities was already doing that in its isIgnored check.  Hopefully that was just half-migrated code and not some cunningly deliberate duplication of effort.  The lifetime of the FileIgnorer is one workspace scan.  That means that rescanning the workspace will reload the locally ignored directories pattern when I've added that feature to FileIgnorer.  That will save me an extra Edit restart every few months.