bin/terminator: The one day, Rodney, answer to the TODO:
the binary distributions won't contain the source, and will be pre-built anyway. the source distributions won't have anything compiled, and you'll have to build it yourself.

A more pragmatic reason for doing the same:

(21:49:05) hughes_elliott@hotmail.com: as far as we know, the attempted bootstrap is no use to anyone. especially because it won't help in the potentially interesting cases like "wrong libc version" or "64- instead of 32-bit".

And, expanding on the same:

There are users who install-everything (which does this for them, falling back to trying to download a native-dist) and their are Subversion users, who build their own anyway.