bin/install-everything.sh: All of our BlueArc users now have the latest JDK installed.  So they can all build the Java.  That will be useful because it means that we'll be able to remove the classes from the dist.  That will make the dist smaller, so using less bandwidth.  It will also make it closer to how we'd like it to be in the longer term: when we have a binary installer for all supported platforms.
This version of install-everything.sh needs to be on all of the affected users' machines before we remove classes from the dist.  I think some of the users may only be running it weekly.  The consequence of it not being on the machines is a torrent of errors starting with:

Recompiling the world...
warning: [path] bad path element "/usr/local/www.jessies.org/terminator/../salma-hayek//classes": no such file or directory
/usr/local/www.jessies.org/terminator/src/terminator/model/TextBuffer.java:7: package e.util does not exist
import e.util.*;
^