dist-exclude: Exclude the classes directory from the dist.
universal.make: This removes the need to build the Java before making the dist.

Elliott's argument here was, as before, that a complete binary distribution of the kind we want to work towards will include the .class files (or .jar file) anyway, so you won't want these wasting bandwidth and space in the source dist.  Also, as before, all our current users have an up-to-date install-everything.sh (which now builds the Java) or build themselves from Subversion.