src/e/edit/Edit.java: fix the code that's supposed to open files in external applications to cope better with Linux, to handle the output streams of the process (this might have caused the program to block if it produced enough output), and to move the names of the various platforms' "open" commands into the code.
edit.properties-sample: remove the properties for the Mac OS and Win32 "open" commands, and extend the list of extensions that should be opened in external applications so I could test this. I wonder if all binary files should be handed over? Or all binary files should cause a dialog to ask what you want to do? Or we should have a built-in hex editor?