The BlueArc echo-local-non-source-directory-pattern script is in a different, site-wide repository.  One of the requirements for finding the ignored-directory-pattern configuration was that it shouldn't necessarily require check-in privileges to the repositories to which it applies.  Another was that it should "just work" for a BlueArc user, given we can make arbitrary hacks to their environment and add scripts to their PATH.  I tried to be clear about that in choosing a name for the script but I'm not particularly impressed by the resulting name and it obviously didn't have the desired effect.
src/e/edit/FileIgnorer.java: Revert to searching for the script on the PATH.

bin/echo-local-non-source-directory-pattern: Provide a default implementation.  Thanks to the change I just made to salma-hayek, this will be on the PATH, but lower down than the BlueArc script.  That feels too fragile to me.  Writing this comment, I see we have a FileUtilities.findOnPath which we could use instead of the default script.  Or we could just hide the errors.