src/terminator/Options.java: Having error-exit-holding always on or always off is never what any user wants.  So remove the misleading, non-useful piece of configuration.
src/terminator/terminal/TerminalControl.java: 
src/terminator/view/JTerminalPane.java: And put the code back essentially how Phil originally wrote it.

Elliott's broken this code twice.  Neither time did he give a justification in the code or check-in comment.  Well, that sometimes happens through the best of intentions when busy, for example, fixing the preferences code.  However, he's never given a satisfactory explanation and there's no sign that he's going to, despite repeated requests and my holding off on this reversion for over a week.  My conclusion is that he hasn't thought about it clearly enough.

This was what he has said, with my commentary interspersed [

i was always uncomfortable with that because it's only a heuristic.

].  The only, very rare, instance that's been presented where the heuristic gets it wrong, the preference gets it wrong too.  That's where you've started a command which succeeded but produced interesting output in a new terminator.  What's required to solve that is an additional command line "+hold" switch, to use the xterm terminology.

Elliott [

if we had a way to know that the just-exited process was bash, i'd be happy.

].  That's irrelevant because, as I said in the now-reinstated code comment, whatever the shell, the user will have seen any failure in the shell window.  We're not talking about a case where the shell has crashed here.  We're under a process.didExitNormally guard.

That reminds me: when he first added the preference, under the comment [

fix the errorExitHolding behavior by removing cruft.

], we weren't using the JNI child launcher.  We were checking the returned value from Java's Process.waitFor.  It's quite probable that that did cause us to do the wrong thing if the shell crashed.  It took us some while after the switch to JNI to get the waitpid, WIFEXITED etc incantation right.  Perhaps we don't have it right yet (though I've just tested it and I think we do) and perhaps bash does crash on Mac OS X (that does ring a bell).  If our waitpid code's broken, our waitpid code wants fixing.  We don't need a gruesome hack that's exposed to the user.

Elliott supplies another reason why bash is irrelevant [

it's odd that tcsh(1) has the same exit-mirroring behavior as bash(1)

].  Mirroring the exit status of the last run command will improve the reliability of scripts.  The bash and tcsh authors haven't special-cased interactive shells.  That's not really odd, although it's arguably unfortunate.

This is what I think is distracting Elliott and keeping him from thinking about this clearly [

Apple actually have code in Terminal to see what process is running so
they can distinguish shells from non-shells.

].  We don't need that to distinguish shells from non-shells.  Phil got this right from the beginning.  We're invoked differently to start a shell than we are to start other commands.  "New Shell" or "terminator" without a command argument tells us that the user is starting a shell.

Perhaps, however, some users want to start their shells with command arguments.  Perhaps, today, they're having a tcsh day.  Do we have any users like that?  I doubt it but such users would be in a perfect position to add "-hold" to the arguments.  Until that's implemented, though, they're going to get errorExitHolding in their tcsh windows.

There's no argument with this! [

not using Terminator's functionality is fine. breaking Terminator's
functionality would be wrong.

].  Say the preference was still in the code.  Our peculiar tcsh user wouldn't want to use it because it would mean that their telnet windows would disappear without telling them the error message.  So this change is not the removal of some useful functionality: it's the removal of something which no-one who understood it would ever want to use.  It's getting it right automatically, all of the time, instead of helping the user to manually get it wrong.

More distraction [

they follow the process hierachy too, so vim-in-bash won't close without a fight but bash-at-the-bash-prompt will.

].  That's a neat additional feature.  Let's implement that some time.  It's orthogonal.

Elliott claims (and I don't dispute) that Terminal and xwsh behave like the preference code I'm removing [

the original (and now current) behavior gives it the same interpretation as in other terminal emulators (xwsh and Terminal).

] though he later confuses me with [

i don't think xwsh(1)'s errorExitHolding used to behave like it does on
Mac OS with Terminal and bash(1). (that combination behaves exactly as we
do at the moment.)

].  There's plenty of precedent for closing without formality when bash exits with a failure status: rxvt, xterm, gnome-terminal and konsole all do this.  We don't need to copy options which no-one finds useful.  Anyone who does find them useful should explain why, so we can have a constructive discussion.  I'm not averse to the solution involving some work, as this check-in comment shows.

Elliott has a number of suggestions for coping with the code as it was [

people who don't like it can turn it off.

].  That's just not acceptable because it means that telnet error messages are lost.  Another [

if you don't like errorExitHolding, turn it off in your preferences and
override the preference when you run Terminator from the command-line, as
you suggested.

] by which I presume he means "turn it on each time you run telnet".  That would work but, like the next suggestion, it inconveniences all reasonable users (who have to change a preference) for the benefit of, apparently, no-one [

if you don't like bash's default behavior, you can always put "exit 0" at the end of "~/.bash_logout".

].  I presume that would work but "bash won the shell wars".  Until this discussion started, we didn't let the user use a different shell for their New Shell windows.  It's worth five lines of simple cruft to better support bash, given that every single terminator user uses it as their shell.

I think this is Elliott's way of moaning about the division between TerminalPane and TerminalControl, something he's complained about before [

(i also dislike duplicating state, especially when it's duplicated in the
wrong class.)

].  The Pane/Control thing is probably a good point but orthogonal and, anyway, there's no duplicated state in this patch.  There was duplicated state between the preference code and the bastardized code I came up with when I first "fixed" this problem.  At that time, I didn't ask what purpose was served by the preference code: I just tried to accommodate it.  I didn't look at the history either.