src/terminator/Terminator.java: It would be at least potentially useful for "terminator -n badger" to create a shell window titled "badger".  It doesn't seem useful or consistent to create a UI with no terminal panes.  Other reasonable alternatives would be to generate a syntax error or simply to ignore the name.
I was looking at this because I'm thinking of parsing "-hold" (xterm) or "--noclose" (konsole) here and wondering whether to pass the option to newShell, even though it isn't useful, for consistency.

src/terminator/TerminatorFrame.java: 
src/terminator/TerminatorMenuBar.java: I could use a default argument or (better) separate newShell from newShellWithTitle but I'd rather make it clear that we're choosing not to set the title.  We might want to derive the title from that of the current pane.  Are we calling it "name" or "title" - we're inconsistent.

src/terminator/view/JTerminalPane.java: Allow the title to be customized but keep the old default, in the same style as newCommandWithTitle.