src/terminator/Terminator.java: Parse a --hold argument to allow you, for motivating example, to inspect the output of "terminator --hold env".  I've wanted to do this several times recently and, without this, it's quite hard.  Try to start bash with a -c command argument from terminator and you'll see what I mean.  The only way I'd previously got this to work was "terminator bash -c env&&false" or by writing a script to wrap env and exit with failure.  It's no use starting bash and running "set" or "env" there.  bash's startup scripts usually change the environment significantly.  Today I wanted this to prove that terminator is adding its own bin to the PATH.  BlueArc's startup scripts remove it again (as startup scripts are wont to do).
I *think* that env is the only thing I ever manually start bash for, so I could be convinced that a script is the way to solve this.  This was, however, the only presented case where the errorExitHolding heuristic fails.  This is my going the extra mile to fix that.

src/terminator/terminal/TerminalControl.java: Pass enough information to decide whether to hold on exit.

src/terminator/view/JTerminalPane.java: Store the configuration.  Terminal uses a tri-state option for don't hold, hold on error, always hold.  Embodying that might make the implementation nicer.  Two boolean parameters constitutes a bad sign.