native/all/launcher/launcher.cpp: Remove what looks to be the only piece of debugging output inadvertently checked in.  Noticed while debugging the Cygwin install-from-clean.
universal.make: If we wanted a Ruby/Java-style getAbsolutePath (or realpath) here now, we'd defer to Ruby (now we're happy to rely on 1.8), so let's just do enough for what we need.  $(SALMA_HAYEK) is supposed to be an absolute directory name, not an arbitrary path, so it shouldn't end in a slash.  (And fix JAVAHPP not to treat it as half a path, which makes no sense.)  The slash doubling was confusing the WiX tools.  Note that $(dir) gives you the trailing slash (because doing it like that removes the need for special cases).

Having $(PROJECT_ROOT)/ as the prefix of $(SALMA_HAYEK)/ for non-salma-hayek projects caught me out, so make $(SALMA_HAYEK) closer to a canonical path, and so closer to what I expected.  This caught me out when adding the AppleJavaExtensions and swing-worker jar files to salma-hayek's installer file list.

This fixes the Cygwin install-from-clean.  Another part of the WiX tutorial that I've been ignoring says that you're supposed to specify which files should be removed when the package is deinstalled.  It was hard to believe that the default is "none" rather than "all the files you installed" (with the exception of anything the user's changed - the kind of nice feature you'd expect a proper installer system to give you for free).  Because I haven't done as suggested, I haven't often been properly testing an install-from-clean.