RobP's home directory is on ukbluearc but he runs terminator on robroy.  The hack in install-everything to remove the .terminator-server-port files was only intended to work for users with locally mounted home directories.  It was like that to avoid trying to remove the files from any home directories which happen to be mounted and which might not be currently accessible, especially not by root - the user running install-everything.
bin/terminator: This is a more robust place to do the hack.  Here we're running as the right user and we don't have to grovel around in the mount table to find the files which need removing.  We can easily tell that install-everything.sh (or, possibly, a proper installer) has been run.  The remainder of the gap between the actual test and the desired test only affects people who know about DEBUGGING_TERMINATOR.  Hmm.  We could test the timestamp of something which is changed by every invocation of make.  The main .class file perhaps.

To come: a salma-hayek check-in to remove the install-everything hack.