src/terminator/terminal/PtyProcess.java: expose the direct child's process id and the functionality for listing (on Mac OS) all the processes using the tty.
native/all/pty/terminator_terminal_PtyProcess.cpp: return the empty string on other OSes, and include the pid of each process in the result on Mac OS, initially for the user's benefit (though it turns out to be useful for our currently implementation, too).

src/terminator/terminal/TerminalControl.java: give access to the PtyProcess.

src/terminator/InfoDialog.java: add a field to the "Info" dialog showing the processes using the tty.

src/terminator/TerminatorFrame.java: handle window closing ourselves, so we can go through the terminals confirming closure with the user if need be.

src/terminator/TerminatorMenuBar.java: also divert the "Close" action via our new code that checks with the user. (And, unrelated to the rest of this patch, add a "Debug" menu. I still need to polish this, but it's been useful enough to me already that I'm not going to remove it.)

src/terminator/view/JTerminalPane.java: rename "errorExitHolding" to "wasCreatedAsNewShell", which is what causes it to be set, rather than the behavior that being set used to cause. It's now also used in new code that confirms with the user that they really want to close a terminal that isn't just running the shell they started in it.