universal.make: This is how g++ rejects -std=c99 even when given plain C:
martind@duezer:~/playpen$ g++ -std=c99 -c plain-old-c.c 
cc1plus: warning: command line option "-std=c99" is valid for C/ObjC but not for C++
martind@duezer:~/playpen$

CFLAGS shouldn't have been passed to the C++ compiler.  That was the reason for splitting CFLAGS from C_AND_CXXFLAGS.  They were because of the CC = $(CXX) assignment.  That was done in order to support linking using the default rule for linking % from %.o, which uses $(LINK.o), which defaults to:

LINK.o = $(CC) $(LDFLAGS) $(TARGET_ARCH)

Happily, I got rid of the dependency on this rule already in revision 1027.  I wanted to build launcherw.exe from launcher.o and found that it took less code if I overrode the default rule.

So this change deletes an obsolete hack, which is nice.  However, if we want to support plain C, we should have some checked in or the support will rust again.