On a default Cygwin installation, C:\cygwin\bin isn't on the PATH.  Combined with the terminator change preceding this, this gets terminator working in that situation.
bin/find-jdk-root.rb: Defer to target-os.rb for the uname call in the usual clumsy way.  /usr/bin/which works on duezer (Linux), icarus (Solaris 8), ted (Darwin 8) and mdorey (Cygwin).

bin/invoke-java.rb: cygpath is always going to be in /bin.

bin/target-os.rb: uname is available in /usr/bin on all the above systems except the Linux one and in /bin on all the above systems except the Darwin one.