www/index.html: Fiddle with the make version description so it doesn't distort what I think is the truth:
I've long suspected that make-3.80-9 only fixes some of the fatal problems that we sometimes run into.  Two people had fatal problems on Linux where they might have been using a 3.80 make and which didn't recur with a 3.81beta4 make.  Elliott's never had a fatal problem with 3.80-9 on Mac OS X.  I've never seen a fatal problem that was definitely on a 3.81beta on Linux.