We were linking two different definitions of unix_exception.h with pty.  That's illegal.  The wrong one - the one from salma-hayek - was being using for pty's error reporting.
native/Headers/unix_exception.h: Just define the class, not its methods.

native/all/launcher/unix_exception.cpp: Link the simpler definition of the method with launcher.

per-directory.make: An underscore in the name of a .cpp file is no longer a sufficient indication that it's JNI source.  Go looking for a corresponding .java file instead.  That's more complicated but The Right Way.

Corresponding change to terminator coming up.