native/all/pty/unix_exception.h: Lose the method definition to unix_exception.cpp and the class definition to One True copy in salma-hayek.  The makerules incorrectly do not clean up obsolete links in .generated.  You'll have to make clean to get rid of the link to this file.
native/all/pty/unix_exception.cpp: The errnoToString implementation of unix_exception is the only one which will now be linked with pty.  This fixes a One Definition Rule violation, though the motivation was to fix the error reporting to use this variant again instead of the simple version included by DirectoryIterator.h from salma-hayek.