per-directory.make: Relax the regular expression for converting the WiX definitions file into a references file to cope with trailing GUIDs.
bin/file-list-to-wxi.rb: Defer to an external program to generate the GUIDs for the components on the fly.  I'm using:

cat > /usr/local/bin/uuid << EOF
#!/bin/bash
ssh duezer uuid "\$@"
EOF
chmod a+rx /usr/local/bin/uuid

Where duezer's a Debian box to which I can ssh without formality and on which I've done sudo apt-get install uuid.  That was the least effort hack I found which was suitable for script-driven operation.

This makes an installer which will completely de-install itself, shortcut and all, when asked.  It passes orca's validation suite with nary a chirp.  It will automatically repair if asked to install again when I've deleted half the files.  And, if you change the ProductCode, it'll update over the top of an earlier version (with the same UpgradeCode, I presume) without bleating.