bin/make-mac-os-app.rb: Take the whole version string as an argument.  I guess Elliott was being deliberately consistent with the Windows-imposed version numbering.  If we're going to be consistent, we'd be better off with one copy of the code.  I've tested this still builds but not that it works.
native/Cygwin/WiX/salma-hayek.wxs: So lose the "1.0." here too.

universal.make: And construct the version here.  Elliott noted that we could do with incorporating the salma-hayek revision too and we can easily do that here.  Ah, and the immediate evaluation of REVISION := `svn info` explains why install-everything.sh had been bleating about '.' not being a working copy.  We can trivially fix that while we're here.

Tidy-up mention of terminator's old product GUID and the no-longer necessary experimental installation commands for the Windows installer.  Make it easy for...

per-directory.make: ... the "candle" rule to receive a unique product GUID every time, along with the version string, in a way that we can incorporate into the .wxs file.