bin/which.rb: Cobble together a path search in Ruby from various things on the web, mainly http://facets.rubyforge.org/rdoc/classes/FileUtils.html#M000379.
bin/find-jdk-root.rb: Use this instead of which(1) to find Java, eliminating the need for knowing where which(1) is installed or the need to have it on the path.  (Although we still need Java on the path!)

bin/target-os.rb: Use Elliott's cunning Config::CONFIG replacement for uname.

This should reduce the number of flickering windows on starting terminator on Windows.