native/all/launcher/launcher.cpp: The error message when using a 1.4.2 JRE was essentially:
FindClass("terminator/Terminator") failed

That's not very useful.  We already sanity-check the JRE version string.  Change the sanity check to exclude anything earlier than 1.5 and, at Elliott's suggestion, anything 1.6 or greater.

Then we get the more useful error message which I've updated to indicate the requirement that the JRE/JDK be 1.5.